pyproject.toml
[project]
name = "ainnovators"
version = "0.1.0"
description = "Multi-agent AI system simulating a startup company"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [
    { name = "Sergei Gordeichuk" },
    { name = "Claude Code Opus 4.5"}
]
keywords = ["ai", "agents", "autogen", "startup", "multi-agent"]

dependencies = [
    "autogen-agentchat>=0.4",
    "autogen-ext[anthropic]>=0.4",
    "autogen-ext[openai]>=0.4",
    "python-pptx>=1.0",
    "python-docx>=1.1",
    "reportlab>=4.2",
    "python-dotenv>=1.0",
    "click>=8.1",
    "requests>=2.31",
    "anthropic>=0.75.0",
    "tiktoken>=0.12.0",
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
    "playwright>=1.40",
    "aiohttp>=3.9",
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
ainnovators = "ainnovators.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ainnovators"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
